Changing SingleCellMailer.users_email behavior
The "All Users" email for admins was recently changed to deliver emails in batches of 500.  It was then discovered that blocks in ActionMailer methods that call `mail` only perform deliveries for the last entry in the block.  This update changes that behavior to iterate over a block in a separate method called `batch_users_email` that then calls the `users_email` method to deliver an email to a batch of 500 users.