changing bind mode re: gRPC FUSE changes to Docker CE for development
According to https://github.com/docker/for-mac/issues/1592, using 'delegated' appears to work the same as 'cached', meaning that updates from the host may be delayed in appearing in the container.  This behavior has been verified locally - changes to source (models, controllers) are not reflected in the container unless passenger is bounced, or the container restarted.  This behavior is consistent using both gRPC FUSE and osxfs.  Setting the mode to 'consistent' and using gRPC as the file system driver fixes the issue.