refactoring DelayedJobAccessor into standalone module
Moves DelayedJobAccessor into app/lib and implements as a standalone module, similar to ErrorTracker.  This allows for simpler code and better testability, as model concerns employing class methods cannot be tested directly in the Rails test environment.  The accepted paradigm is to define a dummy class in the test and test that, which seems to defeat the purpose of the concern altogether.